포스트작성 ,알람기능 추가

+++ src/main/java/Prochat_project/repository/MemberRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
     Optional<MemberEntity> findByMemberId(String memberId);
     List<MemberEntity> findByMemberNameLike(String memberName);
 }
